requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
boto==2.49.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
common-packages==1.9
dnspython==2.4.2
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
jeepney==0.8.0
latest-user-agents==0.0.3
marshmallow==3.20.2
MongoDBProxy==0.1.0
MouseInfo==0.1.3
mss==9.0.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
PyAutoGUI==0.9.54
pydantic==2.2.1
pydantic_core==2.6.1
PyGetWindow==0.0.9
pymongo==4.5.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
python3-xlib==0.15
pytweening==1.0.7
pytz==2024.1
requests==2.31.0
schedule==1.2.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.2.1
uvicorn==0.23.2
